several bugs fixed.

+++ app/src/main/java/com/elkcreek/rodneytressler/musicapp/ui/SearchView/SearchPresenter.java
@@ -25,6 +25,8 @@ public class SearchPresenter implements BasePresenter<SearchView> {
     private final RepositoryService repositoryService;
     private SearchView view;
     private CompositeDisposable disposable;
+    private boolean isSearching;
+    private String searchText;
 
     @Inject
     public SearchPresenter(MusicApiService musicApiService, SharedPreferences sharedPreferences, RepositoryService repositoryService) {
@@ -45,8 +47,8 @@ public class SearchPresenter implements BasePresenter<SearchView> {
 
     private Consumer<List<MusicApi.Artist>> updateViewWithTopArtist() {
         return topArtists -> {
-                view.loadArtists(topArtists);
-                view.hideProgressBar();
+            view.loadArtists(topArtists);
+            view.hideProgressBar();
         };
     }
 
@@ -80,19 +82,22 @@ public class SearchPresenter implements BasePresenter<SearchView> {
         int savedWeek = sharedPreferences.getInt(Constants.WEEKOFYEAR, 0);
         int savedYear = sharedPreferences.getInt(Constants.YEAR, 0);
 
-        if (savedYear == 0 || (weekOfYear != savedWeek && year != savedYear)) {
-            repositoryService.deleteTopArtists();
-            disposable.add(repositoryService.getTopArtistsFromNetwork().subscribe(updateViewWithTopArtist(), updateUiWithError()));
-            SharedPreferences.Editor editor = sharedPreferences.edit();
-            editor.putInt(Constants.WEEKOFYEAR, weekOfYear);
-            editor.putInt(Constants.YEAR, year);
-            editor.apply();
+        if (!isSearching) {
+            if (savedYear == 0 || (weekOfYear != savedWeek && year != savedYear)) {
+                repositoryService.deleteTopArtists();
+                disposable.add(repositoryService.getTopArtistsFromNetwork().subscribe(updateViewWithTopArtist(), updateUiWithError()));
+                SharedPreferences.Editor editor = sharedPreferences.edit();
+                editor.putInt(Constants.WEEKOFYEAR, weekOfYear);
+                editor.putInt(Constants.YEAR, year);
+                editor.apply();
+            } else {
+                disposable.add(repositoryService.getTopArtists().subscribe(
+                        updateViewWithTopArtist()
+                ));
+            }
         } else {
-            disposable.add(repositoryService.getTopArtists().subscribe(
-                    updateViewWithTopArtist()
-            ));
+            artistSearchTextChanged(searchText, true);
         }
-
     }
 
     @Override
@@ -101,6 +106,8 @@ public class SearchPresenter implements BasePresenter<SearchView> {
     }
 
     public void artistSearchTextChanged(String artistSearchText, boolean adapterHasItems) {
+        this.isSearching = artistSearchText.length() > 0;
+        this.searchText = artistSearchText;
         if (disposable == null) {
             disposable = new CompositeDisposable();
         }
@@ -131,6 +138,7 @@ public class SearchPresenter implements BasePresenter<SearchView> {
     }
 
     public void onArtistClicked(MusicApi.Artist artist) {
+        view.showProgressBar();
         view.showMainArtistScreen(artist);
     }
 }
